data/requirements.txt
pandas
google-generativeai
google-auth